Fixed some remaining issues.
git-svn-id: https://svnserver.informatik.kit.edu/i43/svn/code/Palladio/Incubation/SimuLizar/trunk/org.palladiosimulator.simulizar@26535 769cc9a0-fb0c-0410-8c2f-a2305f0e6268

+++ src/org/palladiosimulator/simulizar/runtimestate/SimulatedBasicComponentInstance.java
@@ -29,7 +29,7 @@ public class SimulatedBasicComponentInstance extends SimulatedComponentInstance 
         for (PassiveResource passiveResource : passiveResources) {
             final long initialCount = StackContext.evaluateStatic(
                     passiveResource.getCapacity_PassiveResource().getSpecification(), Integer.class,
-                    context.getStack().currentStackFrame()).intValue();
+                    context.getStack().currentStackFrame());
             final IPassiveResource simulatedResource = new SimSimpleFairPassiveResource(passiveResource, myAssCtx,
                     getRuntimeState().getModel(), initialCount);
             this.passiveResourcesMap.put(passiveResource.getId(), simulatedResource);
